+++ examples/data-sources/sse_ips_profile/data-source.tf
@@ -1,0 +1,7 @@
+data "sse_ips_profile" "example" {
+  name = "Standard IPS Profile"
+}
+
+output "ips_profile_id" {
+  value = data.sse_ips_profile.example.id
+}