+++ examples/data-sources/sse_ips_profiles/data-source.tf
@@ -1,0 +1,5 @@
+data "sse_ips_profiles" "all" {}
+
+output "all_ips_profiles" {
+  value = data.sse_ips_profiles.all.ips_profiles
+}