+++ examples/data-sources/sse_security_profiles/data-source.tf
@@ -1,0 +1,6 @@
+# Fetch all security profiles
+data "sse_security_profiles" "all" {}
+
+output "all_profiles" {
+  value = data.sse_security_profiles.all.security_profiles
+}